requirements.txt
imageio
omegaconf
timm==0.6.12
Pillow==8.3.1
tensorboard
torchsummary
einops
gluoncv
scikit-learn
setuptools==59.5.0

